pyproject.toml
[tool.poetry]
name = "alezi06-gendiff"
version = "0.7.0"
description = ""
authors = ["alezi06 <alezi06@mail.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
flake8 = "^3.8.1"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
